[feature #2415] refactor and fix unit tests
