+++ test/aframeViewer.test.js
@@ -1,0 +1,5 @@
+describe('aframeViewer/index.js', () => {
+  it('init', () => {
+    expect(true).toBe(true)
+  })
+})
