+++ test/helpers.test.js
@@ -8,6 +8,6 @@ describe('checkPanoramaFormat', () => {
   it('throws error', () => {
     expect(() => {
       checkPanoramaFormat(panorama)
-    }).toThrowError('panorama panoramaId is required')
+    }).toThrowError(new Error('panoramaId is required'))
   })
 })
